+++ database/migrations/2024_08_29_062055_create_orders_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->text('product_code');
             $table->text('product_category');
             $table->integer('qty');
-            $table->bigInteger('product_price');
+            $table->decimal('product_price',10,2);
             $table->timestamps();
         });
     }
